backend/requirements.txt
Flask==2.0.3
pymongo[srv]==3.12.3
python-dotenv==1.0.1
supertokens_python==0.22.0
